try to run the tunnel in a test, like fully
