+++ rust/connlib/tunnel/src/device_channel.rs
@@ -346,51 +346,3 @@ mod ioctl {
         mtu: libc::c_int,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #[derive(Clone, Default)]
-    struct Callbacks {}
-
-    impl connlib_shared::Callbacks for Callbacks {}
-
-    // I assume this is not practical to run on macOS because of the tight restrictions on NetworkExtensions
-    // It requires sudo on Linux and elevation on Windows, since it creates the tunnel interface
-    #[cfg(target_os = "linux")]
-    #[tokio::test]
-    async fn device_linux() {
-        device_common();
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn device_windows() {
-        // Install wintun so the test can run
-        // CI only needs x86_64 for now
-        let wintun_bytes = include_bytes!("../../../gui-client/wintun/bin/amd64/wintun.dll");
-        let wintun_path = connlib_shared::windows::wintun_dll_path().unwrap();
-        tokio::fs::create_dir_all(wintun_path.parent().unwrap())
-            .await
-            .unwrap();
-        tokio::fs::write(&wintun_path, wintun_bytes).await.unwrap();
-
-        device_common();
-    }
-
-    fn device_common() {
-        let mut dev = super::Device::new();
-
-        let config = connlib_shared::messages::Interface {
-            ipv4: [100, 71, 96, 96].into(),
-            ipv6: [0xfd00, 0x2021, 0x1111, 0x0, 0x0, 0x0, 0x0019, 0x6538].into(),
-            upstream_dns: vec![connlib_shared::messages::DnsServer::IpPort(
-                connlib_shared::messages::IpDnsServer {
-                    address: ([1, 1, 1, 1], 53).into(),
-                },
-            )],
-        };
-        let dns_config = vec![[100, 100, 111, 1].into()];
-        let callbacks = Callbacks::default();
-        dev.set_config(&config, dns_config, &callbacks).unwrap();
-    }
-}
